Add equality and (limited) hashing to `Extension` struct
